third_party/WebKit/LayoutTests/http/tests/devtools/startup/sources/debugger/script-formatter-console-expected.txt

Tests that the script formatting changes console line numbers.


Running: testSetup

Running: testConsoleMessagesForFormattedScripts
script-formatter-console.html:9 
Pre-format row message list:
[8]
script-formatter-con…e.html:formatted:12 
Post-format row message list:
[11]
